+++ src/components/settings/RegisterSettings.vue
@@ -2,13 +2,22 @@
   <div class="flex flex-col items-start p-12">
 
     <div class="divider">{{ $t('registerSettings') }}</div>
-    <div class="flex items-center flex-row gap-2 max-w-full w-full">
-      <span class="font-bold">{{ $t('emailSuffix') }}: </span>
-      <input type="text" placeholder="example: @tikmatrix.com" class="input input-sm grow input-bordered"
-        v-model="settings.email_suffix" />
+    <div class="flex items-center flex-row gap-2 max-w-full w-full mt-2">
+      <div class="tooltip tooltip-right" :data-tip="$t('emailsFileTips')">
+        <span class="font-bold">{{ $t('emailsFile') }}: </span>
+        <font-awesome-icon :icon="['fas', 'circle-exclamation']" />
+      </div>
+      <input type="text" placeholder="example: C:/Users/Administrator/Desktop/emails.txt"
+        class="input input-sm grow input-bordered" v-model="settings.emails_file" />
+      <button class="btn btn-sm btn-info ml-2" @click="selectEmails">{{ $t('select') }}</button>
+    </div>
+    <div class="flex items-center flex-row gap-2 max-w-full w-full mt-2">
+      <span class="font-bold">{{ $t('password') }}: </span>
+      <input class="input input-bordered w-full max-w-xs" :placeholder="$t('passwordTips')" v-model="settings.password">
+      </input>
 
     </div>
-    <div class="flex items-center flex-row gap-2 max-w-full w-full">
+    <div class="flex items-center flex-row gap-2 max-w-full w-full mt-2">
       <span class="font-bold">{{ $t('nicknames') }}: </span>
       <textarea class="textarea textarea-success grow  h-16" :placeholder="$t('nicknamesTips')" autocomplete="off"
         v-model="settings.nicknames"> </textarea>
@@ -86,6 +95,18 @@ export default {
       console.log('Selected file path:', filePath);
       // 将 filePath 用于其他操作
       this.settings.avatars_path = filePath
+    },
+    async selectEmails() {
+      const filePath = await open({
+        multiple: false, // 是否允许多选文件
+        directory: false, // 是否选择目录
+        filters: [ // 文件过滤器
+          { name: 'Email File', extensions: ['txt'] },
+        ]
+      });
+      console.log('Selected file path:', filePath);
+      // 将 filePath 用于其他操作
+      this.settings.emails_file = filePath
     },
     get_settings() {
       this.$service.get_settings().then(res => {