+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # 更新日志 / Change Log
 
 1. 修复已知BUG / Fix known bugs
+2. 增加注册邮箱配置 / Add email configuration
+3. 增加注册密码配置 / Add password configuration